+++ packages/unraid-api-plugin-connect/src/event-handler/mothership.events.ts
@@ -5,7 +5,7 @@ import { PubSub } from 'graphql-subscriptions';
 
 import { MinigraphStatus } from '../config/connect.config.js';
 import { EVENTS, GRAPHQL_PUBSUB_CHANNEL, GRAPHQL_PUBSUB_TOKEN } from '../helper/nest-tokens.js';
-import { TimeoutCheckerJob } from '../job/timeout-checker.job.js';
+import { TimeoutCheckerJob } from '../connection-status/timeout-checker.job.js';
 import { MothershipConnectionService } from '../service/connection.service.js';
 import { MothershipGraphqlClientService } from '../service/graphql.client.js';
 import { MothershipSubscriptionHandler } from '../service/mothership-subscription.handler.js';
