+++ packages/unraid-api-plugin-connect/src/service/mothership-subscription.handler.ts
@@ -14,7 +14,7 @@ import { SEND_REMOTE_QUERY_RESPONSE } from '../graphql/remote-response.js';
 import { parseGraphQLQuery } from '../helper/parse-graphql.js';
 import { MothershipConnectionService } from './connection.service.js';
 import { MothershipGraphqlClientService } from './graphql.client.js';
-import { InternalClientService } from './internal.client.js';
+import { InternalClientService } from '../internal-rpc/internal.client.js';
 
 type SubscriptionProxy = {
     sha256: string;
